Enable build and test for Python 3.11 and 3.12 (#5393)
# Description

The build and test for Python 3.12 was skipped so far for Mage, which
has been enabled.

# How Has This Been Tested?
<!-- Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
-->

- [X] Will check to see if there are any issues through GitHub's build
and test workflow.

# Checklist
- [ ] The PR is tagged with proper labels (bug, enhancement, feature,
documentation)
- [X] I have performed a self-review of my own code
- [ ] I have added unit tests that prove my fix is effective or that my
feature works
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation

cc:
<!-- Optionally mention someone to let them know about this pull request
-->

@wangxiaoyou1993
