+++ mage_ai/tests/data_preparation/test_variable_manager.py
@@ -1,5 +1,3 @@
-from unittest.mock import patch
-
 import pandas as pd
 from pandas.testing import assert_frame_equal
 
@@ -10,12 +8,10 @@ from mage_ai.data_preparation.repo_manager import get_repo_config
 from mage_ai.data_preparation.variable_manager import (
     VariableManager,
     get_global_variable,
-    get_global_variables,
     set_global_variable,
 )
 from mage_ai.settings.repo import set_repo_path
 from mage_ai.tests.base_test import DBTestCase
-from mage_ai.tests.shared.mixins import ProjectPlatformMixin
 
 
 class VariableManagerTest(DBTestCase):
@@ -128,50 +124,3 @@ class VariableManagerTest(DBTestCase):
     #         get_global_variables(None, pipeline=pipeline), pipeline.variables
     #     )
     #     self.assertEqual(get_global_variables(pipeline.uuid), pipeline.variables)
-
-
-class VariableManagerProjectPlatformTests(ProjectPlatformMixin):
-    def test_get_global_variable(self):
-        with patch(
-            "mage_ai.data_preparation.variable_manager.project_platform_activated",
-            lambda: True,
-        ):
-            with patch(
-                "mage_ai.data_preparation.models.pipeline.project_platform_activated",
-                lambda: True,
-            ):
-                for settings in self.repo_paths.values():
-                    pipeline = Pipeline.create(
-                        self.faker.unique.name(),
-                        repo_path=settings["full_path"],
-                    )
-                    value = self.faker.unique.name()
-                    pipeline.variables = dict(mage=value)
-                    pipeline.save()
-
-                    self.assertEqual(get_global_variable(pipeline.uuid, "mage"), value)
-
-    def test_get_global_variables(self):
-        with patch(
-            "mage_ai.data_preparation.variable_manager.project_platform_activated",
-            lambda: True,
-        ):
-            with patch(
-                "mage_ai.data_preparation.models.pipeline.project_platform_activated",
-                lambda: True,
-            ):
-                for settings in self.repo_paths.values():
-                    pipeline = Pipeline.create(
-                        self.faker.unique.name(),
-                        repo_path=settings["full_path"],
-                    )
-                    pipeline.variables = self.faker.unique.name()
-                    pipeline.save()
-
-                    self.assertEqual(
-                        get_global_variables(None, pipeline=pipeline),
-                        pipeline.variables,
-                    )
-                    self.assertEqual(
-                        get_global_variables(pipeline.uuid), pipeline.variables
-                    )
