+++ mage_integrations/requirements.txt
@@ -1,25 +1,33 @@
-attrs==17.4.0
+attrs==17.4.0; python_version < '3.12'
+attrs==24.2.0; python_version >= '3.12'
 azure-storage-blob==12.14.1
 backoff
 clickhouse_sqlalchemy~=0.2.4
-couchbase==4.1.1
+couchbase==4.1.1; python_version < '3.11'
+couchbase==4.3.1; python_version >= '3.11'
 deltalake==0.17.2
 elasticsearch==8.9.0
 facebook_business==17.0.2
 gnupg==2.3.1
-google-ads==22.1.0
-google-analytics-data==0.14.2
+google-ads==22.1.0; python_version < '3.11'
+google-ads==23.0.0; python_version >= '3.11'
+google-analytics-data==0.14.2; python_version < '3.11'
+google-analytics-data==0.15.0; python_version >= '3.11'
 google-api-python-client==2.70.0
 google-cloud-bigquery~=3.0
-google-cloud-bigquery-storage==2.16.2
+google-cloud-bigquery-storage==2.16.2; python_version < '3.11'
+google-cloud-bigquery-storage==2.26.0; python_version >= '3.11'
 google-cloud-storage==2.5.0
 httpx==0.23.1
 httpx[http2]
 kafka-python~=2.0.2
-mysql-connector-python~=8.2.0
+mysql-connector-python~=8.2.0; python_version < '3.11'
+mysql-connector-python~=9.0.0; python_version >= '3.11'
 opensearch-py~=2.0.0
-oracledb==1.3.1
-pendulum==2.1.0
+oracledb==1.3.1; python_version < '3.12'
+oracledb==2.4.1; python_version >= '3.12'
+pendulum==2.1.0; python_version < '3.12'
+pendulum==3.0.0; python_version >= '3.12'
 protobuf>=4.0.0
 psycopg2==2.9.3
 psycopg2-binary==2.9.3
