+++ setup.py
@@ -81,12 +81,17 @@ setuptools.setup(
             'google-cloud-storage==2.5.0',
             'gspread==5.7.2',
         ],
-        'hdf5': ['tables==3.7.0'],
+        'hdf5': [
+            "tables==3.7.0; python_version < '3.11'",
+            "tables==3.10.1; python_version >= '3.11'",
+        ],
         'mysql': [
-            'mysql-connector-python~=8.2.0',
+            "mysql-connector-python~=8.2.0; python_version < '3.11'",
+            "mysql-connector-python~=9.0.0; python_version >= '3.11'",
         ],
         'oracle': [
-            'oracledb==1.3.1',
+            "oracledb==1.3.1; python_version < '3.12'",
+            "oracledb==2.4.1; python_version >= '3.12'",
         ],
         'postgres': [
             'psycopg2==2.9.3',
@@ -119,7 +124,8 @@ setuptools.setup(
             'influxdb_client==1.36.1',
             'kafka-python==2.0.2',
             'nats-py==2.6.0',
-            'nkeys~=0.1.0',
+            "nkeys~=0.1.0; python_version < '3.12'",
+            "nkeys~=0.2.0; python_version >= '3.12'",
             'opensearch-py==2.0.0',
             'pika==1.3.1',
             'pymongo==4.3.3',
@@ -169,7 +175,8 @@ setuptools.setup(
             'langchain_community<0.0.20',
             'ldap3==2.9.1',
             'nats-py==2.6.0',
-            'nkeys~=0.1.0',
+            "nkeys~=0.1.0; python_version < '3.12'",
+            "nkeys~=0.2.0; python_version >= '3.12'",
             'openai==1.39.0',
             'opensearch-py==2.0.0',
             'opentelemetry-api==1.22.0',
@@ -187,7 +194,8 @@ setuptools.setup(
             'pyairtable==2.3.3',
             'pydruid==0.6.5',
             'pymongo==4.3.3',
-            'pyodbc==4.0.35',
+            "pyodbc==4.0.35; python_version < '3.12'",
+            "pyodbc==5.0.1; python_version >= '3.12'",
             'redshift-connector==2.0.915',
             'lxml==4.9.4',
             'requests_aws4auth==1.1.2',
